Revert: symlink /proc/mounts to /etc/mtab
