+++ cmd/tether/ops_linux.go
@@ -18,6 +18,7 @@ import (
 	"errors"
 	"fmt"
 	"io"
+	"net/http"
 	"os"
 	"syscall"
 
@@ -26,6 +27,8 @@ import (
 	"github.com/vmware/vic/pkg/dio"
 	"github.com/vmware/vic/pkg/trace"
 )
+
+const runMountPoint = "/run"
 
 type operations struct {
 	tether.BaseOperations
@@ -71,4 +74,33 @@ func (t *operations) SessionLog(session *tether.SessionConfig) (dio.DynamicMulti
 
 	// use multi-writer so it goes to both screen and session log
 	return dio.MultiWriter(f, os.Stdout), nil
+}
+
+func (t *operations) Setup(sink tether.Config) error {
+	if err := t.BaseOperations.Setup(sink); err != nil {
+		return err
+	}
+
+	// symlink /etc/mtab to /proc/mounts
+	var err error
+	if err = tether.Sys.Syscall.Symlink("/proc/mounts", "/etc/mtab"); err != nil {
+		if errno, ok := err.(syscall.Errno); !ok || errno != syscall.EEXIST {
+			return err
+		}
+	}
+
+	// unmount /run - https://github.com/vmware/vic/issues/1643
+	if err = tether.Sys.Syscall.Unmount(runMountPoint, syscall.MNT_DETACH); err != nil {
+		if errno, ok := err.(syscall.Errno); !ok || errno != syscall.EINVAL {
+			return err
+		}
+	}
+
+	// TODO: enabled for initial dev debugging only
+	log.Info("Launching pprof server on port 6060")
+	go func() {
+		log.Info(http.ListenAndServe("0.0.0.0:6060", nil))
+	}()
+
+	return nil
 }
