+++ cmd/tether/ops_windows.go
@@ -18,6 +18,7 @@ import (
 	"errors"
 	"fmt"
 	"io"
+	"net/http"
 	_ "net/http/pprof"
 	"os"
 	"strings"
@@ -139,4 +140,19 @@ func (t *operations) SessionLog(session *tether.SessionConfig) (dio.DynamicMulti
 
 	// use multi-writer so it goes to both screen and session log
 	return dio.MultiWriter(f, os.Stdout), nil
+}
+
+func (t *operations) Setup(sink tether.Config) error {
+
+	if err := t.BaseOperations.Setup(sink); err != nil {
+		return err
+	}
+
+	// TODO: enabled for initial dev debugging only
+	log.Info("Launching pprof server on port 6060")
+	go func() {
+		log.Info(http.ListenAndServe("0.0.0.0:6060", nil))
+	}()
+
+	return nil
 }
